fix: mpl blank plot, pyqtgraph grid resize; reorganize tests
mpl blank plot: Replace blockSignals approach with _inSetData flag
that suppresses redundant _plotData calls from toolbar signals during
setData(). Simpler and doesn't interfere with signal delivery.

pyqtgraph grid resize: Reset all row/column stretch factors and
minimum height before re-arranging the grid. Fixes plots staying
small after reducing from many subplots to few.

Reorganize tests: Move tests from test_regressions.py into their
proper homes:
- Axis orientation, complex splitting, mpl first-plot -> test_plotting
- Selection buttons -> test_data_selector
- Records counter, dataset refresh -> test_qcodes_data
Delete test_regressions.py.
