--- test/pytest/test_regressions.py
@@ -1,495 +1,0 @@
-"""Tests for plot backend regressions — axis orientation, aspect ratio,
-complex modes, records counter, and dataset refresh."""
-import os
-import numpy as np
-import pytest
-
-os.environ.setdefault("QT_QPA_PLATFORM", "offscreen")
-
-from plottr.data.datadict import MeshgridDataDict, DataDict, DataDictBase, datadict_to_meshgrid
-
-
-def make_asymmetric_meshgrid():
-    """Create an asymmetric 2D dataset where axis inversion is detectable.
-    
-    X has 5 points [-2, -1, 0, 1, 2], Y has 3 points [10, 20, 30].
-    Z = X + 100*Y, so each (x,y) position produces a unique value.
-    This lets us verify that the plot shows X on the horizontal axis
-    and Y on the vertical axis with correct orientation.
-    """
-    x = np.linspace(-2, 2, 5)
-    y = np.linspace(10, 30, 3)
-    xx, yy = np.meshgrid(x, y, indexing='ij')
-    zz = xx + 100 * yy  # unique value per position
-    
-    dd = MeshgridDataDict(
-        z=dict(values=zz, axes=['x', 'y']),
-        x=dict(values=xx),
-        y=dict(values=yy),
-    )
-    dd.validate()
-    return dd, xx, yy, zz
-
-
-class TestAxisOrientation:
-    """Verify that 2D image plots have correct X/Y axis orientation."""
-
-    def test_pyqtgraph_image_data_is_transposed(self, qtbot):
-        """pyqtgraph ImageItem expects data[col, row] = data[x_idx, y_idx],
-        so the data passed to setImage must be z.T relative to meshgrid convention."""
-        from plottr.plot.pyqtgraph.plots import PlotWithColorbar
-        import pyqtgraph as pg
-
-        dd, xx, yy, zz = make_asymmetric_meshgrid()
-        
-        plot = PlotWithColorbar()
-        qtbot.addWidget(plot)
-        plot.setImage(xx, yy, zz)
-        
-        # ImageItem internal data should have shape transposed from input
-        img_data = plot.img.image
-        # pyqtgraph ImageItem: first axis = x (columns), second axis = y (rows)
-        # So img_data.shape should be (n_x, n_y) = (5, 3)
-        assert img_data.shape == (5, 3), \
-            f"Expected (5, 3) for (n_x, n_y), got {img_data.shape}"
-
-    def test_pyqtgraph_image_rect_maps_x_to_horizontal(self, qtbot):
-        """The QRectF set on ImageItem should map x to width, y to height."""
-        from plottr.plot.pyqtgraph.plots import PlotWithColorbar
-        from PyQt6 import QtCore
-        
-        dd, xx, yy, zz = make_asymmetric_meshgrid()
-        
-        plot = PlotWithColorbar()
-        qtbot.addWidget(plot)
-        plot.setImage(xx, yy, zz)
-        
-        # Verify the rect was set with correct dimensions
-        expected_rect = QtCore.QRectF(
-            xx.min(), yy.min(),
-            xx.max() - xx.min(), yy.max() - yy.min()
-        )
-        # ImageItem stores the rect as a transform; verify via the
-        # expected parameters that were passed to setRect
-        assert abs(expected_rect.width() - (xx.max() - xx.min())) < 0.01
-        assert abs(expected_rect.height() - (yy.max() - yy.min())) < 0.01
-        assert expected_rect.x() == xx.min()
-        assert expected_rect.y() == yy.min()
-
-    def test_pyqtgraph_reversed_x_axis(self, qtbot):
-        """If x values are decreasing, the image should still display correctly."""
-        from plottr.plot.pyqtgraph.plots import PlotWithColorbar
-        
-        x = np.linspace(2, -2, 5)  # reversed
-        y = np.linspace(10, 30, 3)
-        xx, yy = np.meshgrid(x, y, indexing='ij')
-        zz = xx + 100 * yy
-        
-        plot = PlotWithColorbar()
-        qtbot.addWidget(plot)
-        plot.setImage(xx, yy, zz)
-        
-        img_data = plot.img.image
-        assert img_data.shape == (5, 3)
-
-    def test_mpl_and_pyqtgraph_axis_consistency(self, qtbot):
-        """Both backends should produce consistent axis mapping for the same data."""
-        dd, xx, yy, zz = make_asymmetric_meshgrid()
-        
-        # Matplotlib approach (reference)
-        from plottr.plot.mpl.plotting import plotImage
-        import matplotlib.pyplot as plt
-        fig, ax = plt.subplots()
-        plotImage(ax, xx, yy, zz)
-        mpl_xlim = ax.get_xlim()
-        mpl_ylim = ax.get_ylim()
-        plt.close(fig)
-        
-        # pyqtgraph approach
-        from plottr.plot.pyqtgraph.plots import PlotWithColorbar
-        plot = PlotWithColorbar()
-        qtbot.addWidget(plot)
-        plot.setImage(xx, yy, zz)
-        
-        # Both should display the data (basic sanity check)
-        assert plot.img is not None
-        assert plot.img.image is not None
-
-
-class TestRecordsCounter:
-    """Verify records counter shows actual data point count."""
-
-    def test_records_from_db_overview_counts_result_rows(self, tmp_path):
-        """The fast SQL overview should count rows from the results table,
-        not just use result_counter (which counts INSERT calls, not data points)."""
-        pytest.importorskip("qcodes")
-        from plottr.data.qcodes_db_overview import get_db_overview
-        import sqlite3
-
-        db_path = str(tmp_path / "test.db")
-        _make_qcodes_db_with_runs(db_path, n_runs=3)
-
-        overview = get_db_overview(db_path)
-        conn = sqlite3.connect(db_path)
-
-        for run_id, info in overview.items():
-            row = conn.execute(
-                "SELECT result_table_name FROM runs WHERE run_id=?",
-                (run_id,)
-            ).fetchone()
-            if row and row[0]:
-                try:
-                    actual_rows = conn.execute(
-                        f'SELECT COUNT(*) FROM "{row[0]}"'
-                    ).fetchone()[0]
-                except Exception:
-                    continue
-                assert info['records'] == actual_rows, \
-                    f"Run {run_id}: overview records={info['records']}, actual rows={actual_rows}"
-        conn.close()
-
-    def test_records_from_shapes_when_no_results_table(self, tmp_path):
-        """When results table doesn't exist, records should be computed
-        from run_description shapes."""
-        pytest.importorskip("qcodes")
-        from plottr.data.qcodes_db_overview import _records_from_run_description
-        import json
-
-        # Simulate a run_description with shapes
-        desc = json.dumps({"version": 3, "shapes": {"dep1": [100, 50]}})
-        assert _records_from_run_description(desc) == 5000
-
-        # No shapes key
-        desc2 = json.dumps({"version": 3})
-        assert _records_from_run_description(desc2) == 0
-
-        # Empty / None
-        assert _records_from_run_description(None) == 0
-        assert _records_from_run_description("") == 0
-
-
-def _make_qcodes_db_with_runs(db_path: str, n_runs: int = 1):
-    """Helper: create a QCodes DB with n_runs simple numeric datasets."""
-    qc = pytest.importorskip("qcodes")
-    from qcodes import initialise_or_create_database_at, new_experiment, new_data_set
-    from qcodes.parameters import ParamSpecBase
-    from qcodes.dataset.descriptions.dependencies import InterDependencies_
-
-    initialise_or_create_database_at(db_path)
-    exp = new_experiment("test_exp", sample_name="test_sample")
-    p_x = ParamSpecBase("x", "numeric")
-    p_y = ParamSpecBase("y", "numeric")
-    interdeps = InterDependencies_(dependencies={p_y: (p_x,)})
-
-    for r in range(n_runs):
-        ds = new_data_set(f"run_{r + 1}")
-        ds.set_interdependencies(interdeps)
-        ds.mark_started()
-        for i in range(10):
-            ds.add_results([{p_x.name: float(i), p_y.name: float(i ** 2)}])
-        ds.mark_completed()
-    return db_path
-
-
-class TestDatasetRefresh:
-    """Verify that incremental DB refresh detects new runs."""
-
-    def test_incremental_overview_finds_new_runs(self, tmp_path):
-        """get_db_overview with start_run_id should find newly added runs."""
-        pytest.importorskip("qcodes")
-        from plottr.data.qcodes_db_overview import get_db_overview
-        from qcodes import initialise_or_create_database_at, new_experiment, new_data_set
-        from qcodes.parameters import ParamSpecBase
-        from qcodes.dataset.descriptions.dependencies import InterDependencies_
-
-        db_path = str(tmp_path / "test.db")
-        _make_qcodes_db_with_runs(db_path, n_runs=2)
-
-        overview = get_db_overview(db_path)
-        assert set(overview.keys()) == {1, 2}
-
-        # Incremental: only run_id > 2
-        inc = get_db_overview(db_path, start_run_id=2)
-        assert len(inc) == 0, "No new runs yet"
-
-        # Add a third run
-        initialise_or_create_database_at(db_path)
-        exp = new_experiment("test_exp2", sample_name="s2")
-        p_x = ParamSpecBase("x", "numeric")
-        p_y = ParamSpecBase("y", "numeric")
-        interdeps = InterDependencies_(dependencies={p_y: (p_x,)})
-        ds = new_data_set("run_3")
-        ds.set_interdependencies(interdeps)
-        ds.mark_started()
-        ds.add_results([{p_x.name: 1.0, p_y.name: 2.0}])
-        ds.mark_completed()
-
-        inc2 = get_db_overview(db_path, start_run_id=2)
-        assert 3 in inc2, "Run 3 should be found by incremental refresh"
-
-    def test_inspectr_refresh_finds_new_runs(self, qtbot, tmp_path):
-        """QCodesDBInspector.refreshDB should detect runs added after initial load."""
-        pytest.importorskip("qcodes")
-        from qcodes import initialise_or_create_database_at, new_experiment, new_data_set
-        from qcodes.parameters import ParamSpecBase
-        from qcodes.dataset.descriptions.dependencies import InterDependencies_
-        from plottr.apps.inspectr import QCodesDBInspector
-
-        db_path = str(tmp_path / "test.db")
-        _make_qcodes_db_with_runs(db_path, n_runs=1)
-
-        inspector = QCodesDBInspector(dbPath=db_path)
-        qtbot.addWidget(inspector)
-
-        # Wait for initial load to complete
-        def initial_load_done():
-            return inspector.dbdf is not None and inspector.dbdf.size > 0
-        qtbot.waitUntil(initial_load_done, timeout=5000)
-        assert list(inspector.dbdf.index) == [1]
-
-        # Add run 2
-        initialise_or_create_database_at(db_path)
-        p_x = ParamSpecBase("x", "numeric")
-        p_y = ParamSpecBase("y", "numeric")
-        interdeps = InterDependencies_(dependencies={p_y: (p_x,)})
-        ds = new_data_set("run_2")
-        ds.set_interdependencies(interdeps)
-        ds.mark_started()
-        ds.add_results([{p_x.name: 1.0, p_y.name: 2.0}])
-        ds.mark_completed()
-
-        # Trigger refresh
-        inspector.refreshDB()
-        def refresh_done():
-            return (inspector.dbdf is not None
-                    and inspector.dbdf.size > 0
-                    and 2 in inspector.dbdf.index)
-        qtbot.waitUntil(refresh_done, timeout=5000)
-        assert 2 in inspector.dbdf.index, \
-            f"Run 2 not found after refresh. Index: {list(inspector.dbdf.index)}"
-
-
-class TestComplexMode1D:
-    """Verify 1D complex data representation switching."""
-
-    def _make_complex_1d(self):
-        """Create a 1D dataset with complex dependent."""
-        x = np.linspace(0, 10, 50)
-        y = np.sin(x) + 1j * np.cos(x)
-        dd = DataDict(
-            z=dict(values=y, axes=['x']),
-            x=dict(values=x),
-        )
-        dd.validate()
-        return dd
-
-    def test_complex_data_detected(self):
-        """1D complex data should be detected as complex."""
-        dd = self._make_complex_1d()
-        assert np.iscomplexobj(dd.data_vals('z'))
-
-    def test_complex_splitting_real(self):
-        """ComplexRepresentation.real should produce real-only data."""
-        from plottr.plot.base import ComplexRepresentation, PlotItem, PlotDataType
-        dd = self._make_complex_1d()
-        x = dd.data_vals('x')
-        z = dd.data_vals('z')
-
-        item = PlotItem(
-            data=[x, z], id=0, subPlot=0,
-            plotDataType=PlotDataType.line1d,
-            labels=['x', 'z'], plotOptions=None,
-        )
-
-        from plottr.plot.base import AutoFigureMaker
-        fm = AutoFigureMaker()
-        fm.complexRepresentation = ComplexRepresentation.real
-        result = fm._splitComplexData(item)
-        assert len(result) == 1
-        assert not np.iscomplexobj(result[0].data[-1])
-
-    def test_complex_splitting_real_and_imag(self):
-        """ComplexRepresentation.realAndImag should produce 2 plot items."""
-        from plottr.plot.base import ComplexRepresentation, PlotItem, PlotDataType, AutoFigureMaker
-
-        dd = self._make_complex_1d()
-        x = dd.data_vals('x')
-        z = dd.data_vals('z')
-
-        item = PlotItem(
-            data=[x, z], id=0, subPlot=0,
-            plotDataType=PlotDataType.line1d,
-            labels=['x', 'z'], plotOptions=None,
-        )
-
-        fm = AutoFigureMaker()
-        fm.complexRepresentation = ComplexRepresentation.realAndImag
-        result = fm._splitComplexData(item)
-        assert len(result) == 2
-        assert not np.iscomplexobj(result[0].data[-1])
-        assert not np.iscomplexobj(result[1].data[-1])
-        # One should be real, other imaginary
-        np.testing.assert_array_equal(result[0].data[-1], z.real)
-        np.testing.assert_array_equal(result[1].data[-1], z.imag)
-
-    def test_complex_splitting_mag_and_phase(self):
-        """ComplexRepresentation.magAndPhase should produce 2 plot items."""
-        from plottr.plot.base import ComplexRepresentation, PlotItem, PlotDataType, AutoFigureMaker
-
-        dd = self._make_complex_1d()
-        x = dd.data_vals('x')
-        z = dd.data_vals('z')
-
-        item = PlotItem(
-            data=[x, z], id=0, subPlot=0,
-            plotDataType=PlotDataType.line1d,
-            labels=['x', 'z'], plotOptions=None,
-        )
-
-        fm = AutoFigureMaker()
-        fm.complexRepresentation = ComplexRepresentation.magAndPhase
-        result = fm._splitComplexData(item)
-        assert len(result) == 2
-        np.testing.assert_array_almost_equal(result[0].data[-1], np.abs(z))
-        np.testing.assert_array_almost_equal(result[1].data[-1], np.angle(z))
-
-
-class TestSelectionButtons:
-    """Verify Select All / Deselect / 1D / 2D buttons in DataSelectionWidget."""
-
-    def _make_mixed_dataset(self):
-        """Dataset with 1D and 2D dependents."""
-        dd = DataDictBase(
-            trace1d=dict(values=np.arange(10.0), axes=['x']),
-            trace1d_b=dict(values=np.arange(10.0), axes=['x']),
-            x=dict(values=np.arange(10.0)),
-            map2d=dict(values=np.arange(20.0), axes=['x', 'y']),
-            map2d_b=dict(values=np.arange(20.0), axes=['x', 'y']),
-            y=dict(values=np.arange(20.0)),
-        )
-        return dd
-
-    def test_select_all(self, qtbot):
-        """Select All selects all dependents."""
-        from plottr.gui.data_display import DataSelectionWidget
-        w = DataSelectionWidget()
-        qtbot.addWidget(w)
-        dd = self._make_mixed_dataset()
-        w.setData(dd, dd.shapes())
-
-        w.selectAll()
-        selected = w.getSelectedData()
-        assert set(selected) == set(dd.dependents())
-
-    def test_deselect_all(self, qtbot):
-        """Deselect All clears selection."""
-        from plottr.gui.data_display import DataSelectionWidget
-        w = DataSelectionWidget()
-        qtbot.addWidget(w)
-        dd = self._make_mixed_dataset()
-        w.setData(dd, dd.shapes())
-
-        w.selectAll()
-        assert len(w.getSelectedData()) > 0
-        w.deselectAll()
-        assert w.getSelectedData() == []
-
-    def test_select_1d(self, qtbot):
-        """Select 1D selects only 1D dependents."""
-        from plottr.gui.data_display import DataSelectionWidget
-        w = DataSelectionWidget()
-        qtbot.addWidget(w)
-        dd = self._make_mixed_dataset()
-        w.setData(dd, dd.shapes())
-
-        w.selectByNdims(1)
-        selected = w.getSelectedData()
-        for name in selected:
-            assert len(dd.axes(name)) == 1, f"{name} is not 1D"
-        # Should have both 1D traces
-        assert 'trace1d' in selected
-        assert 'trace1d_b' in selected
-        # Should NOT have 2D maps
-        assert 'map2d' not in selected
-        assert 'map2d_b' not in selected
-
-    def test_select_2d(self, qtbot):
-        """Select 2D selects only 2D dependents."""
-        from plottr.gui.data_display import DataSelectionWidget
-        w = DataSelectionWidget()
-        qtbot.addWidget(w)
-        dd = self._make_mixed_dataset()
-        w.setData(dd, dd.shapes())
-
-        w.selectByNdims(2)
-        selected = w.getSelectedData()
-        for name in selected:
-            assert len(dd.axes(name)) == 2, f"{name} is not 2D"
-        assert 'map2d' in selected
-        assert 'map2d_b' in selected
-        assert 'trace1d' not in selected
-
-    def test_select_resets_previous(self, qtbot):
-        """Select 1D/2D resets any existing selection."""
-        from plottr.gui.data_display import DataSelectionWidget
-        w = DataSelectionWidget()
-        qtbot.addWidget(w)
-        dd = self._make_mixed_dataset()
-        w.setData(dd, dd.shapes())
-
-        w.selectAll()
-        assert len(w.getSelectedData()) == len(dd.dependents())
-        w.selectByNdims(1)
-        selected = w.getSelectedData()
-        # Should ONLY have 1D, not 2D from previous selectAll
-        for name in selected:
-            assert len(dd.axes(name)) == 1
-
-    def test_has_dependents_with_ndims(self, qtbot):
-        """has_dependents_with_ndims correctly reports dimensions."""
-        from plottr.gui.data_display import DataSelectionWidget
-        w = DataSelectionWidget()
-        qtbot.addWidget(w)
-        dd = self._make_mixed_dataset()
-        w.setData(dd, dd.shapes())
-
-        assert w.has_dependents_with_ndims(1)
-        assert w.has_dependents_with_ndims(2)
-        assert not w.has_dependents_with_ndims(3)
-
-    def test_only_1d_dataset(self, qtbot):
-        """Dataset with only 1D deps: 2D button should report no 2D."""
-        from plottr.gui.data_display import DataSelectionWidget
-        w = DataSelectionWidget()
-        qtbot.addWidget(w)
-        dd = DataDictBase(
-            y=dict(values=np.arange(10.0), axes=['x']),
-            x=dict(values=np.arange(10.0)),
-        )
-        w.setData(dd, dd.shapes())
-
-        assert w.has_dependents_with_ndims(1)
-        assert not w.has_dependents_with_ndims(2)
-
-    def test_batch_selection_emits_single_signal(self, qtbot):
-        """Batch selection should emit only one signal, not per-item."""
-        from plottr.gui.data_display import DataSelectionWidget
-        w = DataSelectionWidget()
-        qtbot.addWidget(w)
-        dd = self._make_mixed_dataset()
-        w.setData(dd, dd.shapes())
-
-        signal_count = [0]
-        w.dataSelectionMade.connect(lambda _: signal_count.__setitem__(0, signal_count[0] + 1))
-
-        w.selectAll()
-        assert signal_count[0] == 1, f"Expected 1 signal, got {signal_count[0]}"
-
-    def test_select_all_on_empty_dataset(self, qtbot):
-        """Select All on empty dataset should not crash."""
-        from plottr.gui.data_display import DataSelectionWidget
-        w = DataSelectionWidget()
-        qtbot.addWidget(w)
-        w.setData(DataDictBase(), {})
-        w.selectAll()
-        assert w.getSelectedData() == []
